增加功能: 检测文件数量变化

+++ deskcopy.py
@@ -7,21 +7,21 @@ TARGET = 'D:\\desktop'
 import os,time,shutil,sys
 
 os.chdir(os.path.join(os.path.expanduser('~'),'Desktop'))
-filelst = os.listdir()
-"""
-def monitor_file_number_change():
+filenum = len(os.listdir())
+
+def is_file_number_change():
     '''检测是否出现新文件
-返回值:当前（改变后的）文件列表(list)'''
-    #技术细节：产生一个阻塞，获取桌面文件列表，统计文件数量，和原来的数量对比。
-    #        若增多，说明有新文件，则退出阻塞，进行下一步操作。
-    global filelst
-    while True:
-        nowfilelst = os.listdir()
-        if len(nowfilelst) > len(filelst):
-            yield nowfilelst
-            filelst = nowfilelst    #复制结束后更新“原文件列表”
-        else:
-            time.sleep(60)
+返回值:当前（改变后的）文件列表(list)
+    技术细节：获取桌面文件列表，统计文件数量，和原来的数量对比。
+            若增多，说明有新文件，则退出阻塞，进行下一步操作。'''
+    global filenum
+    nowfilenum = len(os.listdir())
+    if nowfilenum > filenum:
+        filenum = nowfilenum    #复制结束后更新“原文件数量”
+        return True
+    else:
+        return False
+"""
 def get_new_file_name_list(new:list):
     '''获取新增文件的文件名的列表
 new(list):当前（改变后的）文件列表
@@ -41,12 +41,13 @@ def main():
 """
 def main():
     while True:
-        filelst = os.listdir()
-        for i in filelst:
-            target = os.path.join(TARGET,i)
-            shutil.copy(i,target)
-        print(f'已复制 at {time.time()}')
-        time.sleep(3*60 if len(sys.argv) == 1 else 5)    #如果传入参数就是调试模式，缩短间隔时间
+        if is_file_number_change():
+            filelst = os.listdir()
+            for i in filelst:
+                target = os.path.join(TARGET,i)
+                shutil.copy(i,target)
+            print(f'已复制 at {time.time()}')
+        time.sleep(30 if len(sys.argv) == 1 else 1)    #如果传入参数就是调试模式，缩短间隔时间
 
 if __name__ == '__main__':
     main()
